Time: 19 ms (86.68%) | Memory: 96 MB (16.06%) - LeetSync

+++ 4-median-of-two-sorted-arrays/median-of-two-sorted-arrays.cpp
@@ -15,7 +15,7 @@ public:
             return result;
         }
         else{
-            double result=(double)answer[(n/2)+0.5];
+            int result=answer[(n/2)+0.5];
             return result;
         }
         return -1;
